minor tweaks in migration files

+++ database/migrations/2018_01_19_213040_create_cities_table.php
@@ -22,7 +22,7 @@ class CreateCitiesTable extends Migration
             $table->unsignedBigInteger('population')->nullable();
             $table->decimal('latitude', 11, 8)->nullable();
             $table->decimal('longitude', 11, 8)->nullable();
-            $table->unsignedInteger('order_column')->nullable();
+            $table->unsignedBigInteger('order_column')->nullable();
             $table->unsignedTinyInteger('status')->default(1)->comment('0:incomplete, 1:draft, 2:published, 3:Inactive, 4..n:CUSTOM');
             $table->timestamps();
