+++ database/migrations/2021_01_08_003834_create_media_table.php
@@ -23,7 +23,7 @@ class CreateMediaTable extends Migration
             $table->json('custom_properties');
             $table->json('generated_conversions');
             $table->json('responsive_images');
-            $table->unsignedInteger('order_column')->nullable();
+            $table->unsignedBigInteger('order_column')->nullable();
             $table->nullableTimestamps();
         });
     }
